Implement GetDeploymentByName(string)
This is a helper function which acts like `GetDeployment(string)` but takes
a `Deployment.Name` instead of `Deployment.ID`. It is useful for us because
we don't keep track of the deployment ID and instead set the Name to a UUID
that we already know.

This code has been extracted from our Cloud Foundry broker because we feel
like it may be useful to other people and belongs in the library:

- https://github.com/alphagov/paas-compose-broker/blob/da7c995a49fe1d33a4c8062164c1bd147ff2d538/broker/utils.go#L13-L26

In the longer term it would be preferable if the API could provide a filter
for getting deployments so that you didn't need to fetch all and then
iterate over them. I've raised a support ticket (#384655082).