Implement GetClusterByName(string)
This implements a helper function that returns a cluster based on
a name.
It's useful as all calls to enterprise accounts must contain cluster ID.
The only way to do it right now is to do end software based filtering from the output of GetClusters().
This commit moves the functionality to the API library but it would be good to
have such thing in the API to save bandwidth on both sides.