Expose the gorequest logging interface to gocomposeapi consumers.
Without this change it was possible to enable debugging mode on
gorequest, but not to redirect it from os.Stderr to another
destination.

Because the boilerplate for every gorequest to enable this
new logging is so high there is now a `client.newRequest()`
method to hide a great deal of it.